// WebJet/AngularUI/src/app/models/movieDetails.ts
export class MovieDetails {
    Title: string;
    Year: string;
    Rated: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    Writer: string;
    Actors: string;
    Plot: string;
    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Metascore: string;
    Rating: string;
    Votes: string;
    ID: string;
    Price: string;
    Type: string;
    BadImage: boolean;
    MovieServiceTitle: string;
}